I've made troubleshooting changes
I've made sure that values that need to be integers are integers.
I've commented out x0 since it is not used for pso.
I've added a print line for each tested parameter set in the pso.
I've changed the Gillespie step to be shorter than a week i.e. the calculated gillespie step or 1 week (in tdate time).